Separate validation and request building
Make standalone, testable modules.
