+++ lib/barion.js
@@ -7,11 +7,8 @@ const {
     BankTransfer,
     SendTransfer
 } = require('./domain');
-const {
-    buildRequest,
-    buildRequestWithoutValidation,
-    sanitizeThenValidate
-} = require('./request-builder');
+const { buildRequest, buildRequestWithoutValidation } = require('./build');
+const { sanitizeThenValidate } = require('./validate');
 const services = require('./services');
 
 /**
