+++ lib/build.js
@@ -1,6 +1,21 @@
 const { immutableFields } = require('./constants');
-const { BarionModelError } = require('./errors');
-const Joi = require('@hapi/joi');
+const { sanitizeThenValidate } = require('./validate');
+
+/**
+ * Does a case-insensitive search. Returns if the value is presented in the given array.
+ * @param {String[]} array - The set of strings to search between.
+ * @param {String} value - The value to search for. 
+ */
+function containsCaseInsensitive (array, value) {
+    if (!Array.isArray(array) || array.length === 0) {
+        return false;
+    }
+
+    const regexp = new RegExp(value, 'i'); // case-insensitive match
+    const matches = array.filter(value => regexp.test(value));
+
+    return Boolean(matches.length);
+}
 
 /**
  * Returns intersection of two string arrays. The comparison is case insensitive.
@@ -23,30 +38,6 @@ function getCaseInsensitiveIntersection (standardKeys, untrustedKeys) {
     });
 
     return intersection;
-}
-
-/**
- * Picks off error messages from Joi's validation error object to an array.
- * @param {Object} details - The validation error object. 
- */
-function marshalValidationError (error) {
-    return error.details.map(error => error.message);
-}
-
-/**
- * Does a case-insensitive search. Returns if the value is presented in the given array.
- * @param {String[]} array - The set of strings to search between.
- * @param {String} value - The value to search for. 
- */
-function containsCaseInsensitive (array, value) {
-    if (!Array.isArray(array) || array.length === 0) {
-        return false;
-    }
-
-    const regexp = new RegExp(value, 'i'); // case-insensitive match
-    const matches = array.filter(value => regexp.test(value));
-
-    return Boolean(matches.length);
 }
 
 /**
@@ -76,21 +67,6 @@ function buildRequestWithoutValidation (schema, defaults, customs) {
 }
 
 /**
- * Sanitizes, then validates request object with the given schema.
- * @param {Object} schema - Schema, which defines structure of the request object.
- * @param {Object} object - The request object, that is expected to send to API. 
- */
-function sanitizeThenValidate (schema, object) {
-    const { error, value } = Joi.validate(object, schema, { stripUnknown: true, abortEarly: false });
-
-    if (error) {
-        throw new BarionModelError('The given object is invalid.', marshalValidationError(error));
-    }
-
-    return value;
-}
-
-/**
  * Builds request object, that could send to the Barion API.
  * @param {Object} schema - Schema, which defines structure of the request object.
  * @param {Object} defaults - Default values from Barion instance.
@@ -109,6 +85,5 @@ function buildRequest (schema, defaults, options) {
 
 module.exports = {
     buildRequest,
-    buildRequestWithoutValidation,
-    sanitizeThenValidate
+    buildRequestWithoutValidation
 };
